refactor + improvements
