+++ src/components/TokenDistributionChart/index.tsx
@@ -1,54 +1,21 @@
 import { ChartBarSquareIcon, ChartPieIcon } from '@heroicons/react/24/outline'
 import Loading from 'components/Loading'
-import { BigNumber, constants } from 'ethers'
-import {
-  OrderDirection,
-  Participant_OrderBy,
-  ParticipantsDocument,
-  ParticipantsQuery,
-  QueryParticipantsArgs,
-} from 'generated/graphql'
-import { client } from 'lib/apollo/client'
-import { paginateDepleteQuery } from 'lib/apollo/paginateDepleteQuery'
-import { PV } from 'models/pv'
+import { BigNumber } from 'ethers'
+import { ParticipantsQuery } from 'generated/graphql'
 import { useState } from 'react'
-import { useQuery } from 'react-query'
 import TokenAreaChart from './TokenAreaChart'
 import TokenPieChart from './TokenPieChart'
 
 export default function TokenDistributionChart({
-  projectId,
-  pv,
+  participants,
+  isLoading,
   tokenSupply,
 }: {
-  projectId: number | undefined
-  pv: PV | undefined
+  participants: ParticipantsQuery['participants'] | undefined
+  isLoading?: boolean
   tokenSupply: BigNumber | undefined
 }) {
   const [viewMode, setViewMode] = useState<'pie' | 'area'>('pie')
-
-  const { data: allParticipants, isLoading } = useQuery(
-    [`token-holders-${projectId}-${pv}`],
-    () =>
-      paginateDepleteQuery<ParticipantsQuery, QueryParticipantsArgs>({
-        client,
-        document: ParticipantsDocument,
-        variables: {
-          orderDirection: OrderDirection.desc,
-          orderBy: Participant_OrderBy.balance,
-          where: {
-            projectId,
-            pv,
-            balance_gt: BigNumber.from(0),
-            wallet_not: constants.AddressZero,
-          },
-        },
-      }),
-    {
-      staleTime: 5 * 60 * 1000, // 5 min
-      enabled: !!(projectId && pv),
-    },
-  )
 
   // Don't render chart for projects with no token supply
   if (tokenSupply?.eq(0)) return null
@@ -68,7 +35,7 @@ export default function TokenDistributionChart({
     case 'pie':
       content = (
         <TokenPieChart
-          participants={allParticipants}
+          participants={participants}
           tokenSupply={tokenSupply}
           size={size}
         />
@@ -76,38 +43,31 @@ export default function TokenDistributionChart({
       break
     case 'area':
       content = (
-        <TokenAreaChart
-          participants={allParticipants}
-          tokenSupply={tokenSupply}
-          height={size}
-        />
+        <TokenAreaChart participants={participants} tokenSupply={tokenSupply} />
       )
       break
   }
 
   return (
     <div className="w-full">
-      <div
-        style={{ height: size, width: '100%' }}
-        className="flex items-center justify-center"
-      >
-        {content}
-      </div>
+      <div style={{ height: size, width: '100%' }}>{content}</div>
 
-      <div className="mt-4 flex justify-center gap-3">
-        <div
-          onClick={() => setViewMode('pie')}
-          className={viewMode === 'pie' ? 'opacity-100' : 'opacity-50'}
-        >
-          <ChartPieIcon className="h-5 w-5" />
+      {participants?.length && participants.length > 1 && (
+        <div className="mt-4 flex justify-center gap-3">
+          <div
+            onClick={() => setViewMode('pie')}
+            className={viewMode === 'pie' ? 'opacity-100' : 'opacity-50'}
+          >
+            <ChartPieIcon className="h-5 w-5" />
+          </div>
+          <div
+            onClick={() => setViewMode('area')}
+            className={viewMode === 'area' ? 'opacity-100' : 'opacity-50'}
+          >
+            <ChartBarSquareIcon className="h-5 w-5" />
+          </div>
         </div>
-        <div
-          onClick={() => setViewMode('area')}
-          className={viewMode === 'area' ? 'opacity-100' : 'opacity-50'}
-        >
-          <ChartBarSquareIcon className="h-5 w-5" />
-        </div>
-      </div>
+      )}
     </div>
   )
 }
